- fps fixed float display -TranDanh

+++ trunk/SolutionContraGame/ContraGame/CTimeSpanDx9.h
@@ -17,7 +17,7 @@ public:
 	
 	CTimeSpanDx9(UINT64 Ticks);
 	
-	int	getMilliseconds();
+	float	getMilliseconds();
 	
 	UINT64	getTicks();
